cheri: Correctly use uintptr_t in mm/
Correctly use uintptr_t in mm. Add cast operators
as neccessary.

+++ mm/memory-failure.c
@@ -2718,7 +2718,7 @@ static int soft_offline_in_use_page(struct page *page)
 
 	if (mf_isolate_folio(folio, &pagelist)) {
 		ret = migrate_pages(&pagelist, alloc_migration_target, NULL,
-			(unsigned long)&mtc, MIGRATE_SYNC, MR_MEMORY_FAILURE, NULL);
+			(uintptr_t)&mtc, MIGRATE_SYNC, MR_MEMORY_FAILURE, NULL);
 		if (!ret) {
 			bool release = !huge;
